requirements.txt
Flask==3.0.0
gunicorn==21.2.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
packaging==23.2
pyTelegramBotAPI==4.14.0
python-dotenv==1.0.0
requests==2.31.0
urllib3==2.1.0
Werkzeug==3.0.1
aiohttp==3.9.1
aiosignal==1.3.1
async-timeout==4.0.3